+++ client/src/utils/API.js
@@ -4,7 +4,7 @@ export default {
     getProducts: function() {
         return axios.get("/api/products");
       },
-    getWishList: function(id) {
-      return axios.get("/api/wishlist/" + id);
+    getWishList: function() {
+      return axios.get("/api/wishlist/");
     }
 }
